Move Pokemon stat helpers to new module
