+++ utils/sheet_display.py
@@ -3,7 +3,7 @@
 from evennia.utils.evtable import EvTable
 
 from utils.ansi import ansi
-from commands.command import get_max_hp, get_stats
+from pokemon.utils.pokemon_helpers import get_max_hp, get_stats
 from utils.xp_utils import get_display_xp, get_next_level_xp
 from pokemon.stats import level_for_exp
 from utils.faction_utils import get_faction_and_rank
